+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## 2020-08-24
+
+### Added
+
+- Add weighting to dataset search
+- Allow for searching of datasets by source tag name
+
 ## 2020-08-20
 
 ### Added